Kill ViewStateSupport mixin
